+++ src/scripts/plugin.ts
@@ -1,6 +1,6 @@
 import { PiralPlugin } from 'piral-core';
-import { includeSearchProvider } from './searchProviders';
-import { appendSection } from './sitemap';
+import { excludeSearchProvider, includeSearchProvider } from './searchProviders';
+import { appendSection, removeSection } from './sitemap';
 import { PiletDocletApi } from './types';
 
 export function createDocletPlugin(): PiralPlugin<PiletDocletApi> {
@@ -15,8 +15,21 @@ export function createDocletPlugin(): PiralPlugin<PiletDocletApi> {
         },
       }));
     },
+    unregisterDocumentation(section, category) {
+      const routes = removeSection(section, category);
+      context.dispatch((state) => ({
+        ...state,
+        routes: Object.keys(routes).reduce((oldRoutes, path) => {
+          delete oldRoutes[path];
+          return oldRoutes;
+        }, state.routes),
+      }));
+    },
     registerSearchProvider(cb) {
       includeSearchProvider(cb);
+    },
+    unregisterSearchProvider(cb) {
+      excludeSearchProvider(cb);
     },
   });
 }